+++ CHANGELOG.md
@@ -3,5 +3,6 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+* Option to set a timeout on QRadar.fetch, after which an error is thrown. Defaults to 5 seconds (5000 ms).
 ### Changed
 ### Removed